2FA support and error handling stuff

+++ ReArc.Shared/Records/Database/ArcUser.cs
@@ -10,6 +10,6 @@ public record class ArcUser(
     string _id,
 #pragma warning restore IDE1006
     string Email,
-    int StorageSize,
+    long StorageSize,
     int AccountNumber
 );
